# apps/manager/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@sysarmor.com
    name: SysArmor Team
    url: https://sysarmor.com/support
  description: SysArmor EDR 系统的控制平面服务 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sysarmor.com/terms
  title: SysArmor Manager API
  version: "1.0"
paths:
...
  /services/flink/config:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群的配置信息
      produces:
      - application/json
      responses:
        "200":
          description: Flink 配置信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Flink 配置
      tags:
      - flink
...
  /services/flink/jobs:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群中所有作业的信息
      produces:
      - application/json
      responses:
        "200":
          description: 作业列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Flink 作业列表
      tags:
      - flink
  /services/flink/jobs/{job_id}:
    get:
      consumes:
      - application/json
      description: 获取指定作业的详细信息，包括顶点、状态、执行计划等
      parameters:
      - description: 作业 ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 作业详细信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 作业不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取特定作业的详细信息
      tags:
      - flink
...
  /services/flink/jobs/overview:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群中所有作业的概览信息，包括状态统计、分类等
      produces:
      - application/json
      responses:
        "200":
          description: 作业概览信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取作业概览信息
      tags:
      - flink
  /services/flink/overview:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群的基本信息，包括状态、TaskManager 数量、作业数量等
      produces:
      - application/json
      responses:
        "200":
          description: 集群概览信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Flink 集群概览
      tags:
      - flink
  /services/flink/taskmanagers:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群中所有 TaskManager 的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: TaskManager 信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 TaskManager 信息
      tags:
      - flink
  /services/flink/taskmanagers/overview:
    get:
      consumes:
      - application/json
      description: 获取 Flink 集群中所有 TaskManager 的概览信息，包括资源使用统计
      produces:
      - application/json
      responses:
        "200":
          description: TaskManager 概览信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 TaskManager 概览信息
      tags:
      - flink
  /services/flink/test-connection:
    get:
      consumes:
      - application/json
      description: 测试与 Flink 集群的连接状态
      produces:
      - application/json
      responses:
        "200":
          description: 连接成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 连接失败
          schema:
            additionalProperties: true
            type: object
      summary: 测试 Flink 连接
      tags:
      - flink
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"